+++ features/cerberusweb.core/api/uri/profiles/calendar_recurring_profile.php
@@ -195,7 +195,9 @@ class PageSection_ProfilesCalendarRecurringProfile extends Extension_PageSection
 		
 		try {
 			if(!empty($id) && !empty($do_delete)) { // Delete
-				// [TODO] ACL
+				if(!$active_worker->hasPriv(sprintf("contexts.%s.delete", CerberusContexts::CONTEXT_CALENDAR_EVENT_RECURRING)))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.delete'));
+				
 				DAO_CalendarRecurringProfile::delete($id);
 				
 				echo json_encode(array(
@@ -207,13 +209,10 @@ class PageSection_ProfilesCalendarRecurringProfile extends Extension_PageSection
 				return;
 				
 			} else {
-				if(empty($event_name))
-					throw new Exception_DevblocksAjaxValidationError("The 'Name:' field is required.", 'event_name');
-				
-				if(empty($calendar_id))
-					throw new Exception_DevblocksAjaxValidationError("The 'Calendar:' field is required.", 'calendar_id');
-				
 				if(empty($id)) { // New
+					if(!$active_worker->hasPriv(sprintf("contexts.%s.create", CerberusContexts::CONTEXT_CALENDAR_EVENT_RECURRING)))
+						throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.create'));
+					
 					$fields = array(
 						DAO_CalendarRecurringProfile::CALENDAR_ID => $calendar_id,
 						DAO_CalendarRecurringProfile::EVENT_NAME => $event_name,
@@ -226,6 +225,9 @@ class PageSection_ProfilesCalendarRecurringProfile extends Extension_PageSection
 						DAO_CalendarRecurringProfile::PATTERNS => $patterns,
 					);
 					
+					if(!DAO_CalendarRecurringProfile::validate($fields, $error))
+						throw new Exception_DevblocksAjaxValidationError($error);
+					
 					if(false == ($id = DAO_CalendarRecurringProfile::create($fields)))
 						return false;
 					
@@ -233,6 +235,9 @@ class PageSection_ProfilesCalendarRecurringProfile extends Extension_PageSection
 						C4_AbstractView::setMarqueeContextCreated($view_id, CerberusContexts::CONTEXT_CALENDAR_EVENT_RECURRING, $id);
 					
 				} else { // Edit
+					if(!$active_worker->hasPriv(sprintf("contexts.%s.update", CerberusContexts::CONTEXT_CALENDAR_EVENT_RECURRING)))
+						throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.edit'));
+					
 					$fields = array(
 						DAO_CalendarRecurringProfile::EVENT_NAME => $event_name,
 						DAO_CalendarRecurringProfile::EVENT_START => $event_start ?: 'midnight',
@@ -243,10 +248,13 @@ class PageSection_ProfilesCalendarRecurringProfile extends Extension_PageSection
 						DAO_CalendarRecurringProfile::IS_AVAILABLE => $is_available ? 1 : 0,
 						DAO_CalendarRecurringProfile::PATTERNS => $patterns,
 					);
-					DAO_CalendarRecurringProfile::update($id, $fields);
 					
+					if(!DAO_CalendarRecurringProfile::validate($fields, $error, $id))
+						throw new Exception_DevblocksAjaxValidationError($error);
+					
+					DAO_CalendarRecurringProfile::update($id, $fields);
 				}
-	
+				
 				// Custom fields
 				@$field_ids = DevblocksPlatform::importGPC($_REQUEST['field_ids'], 'array', array());
 				DAO_CustomFieldValue::handleFormPost(CerberusContexts::CONTEXT_CALENDAR_EVENT_RECURRING, $id, $field_ids);
