+++ features/cerberusweb.core/api/uri/profiles/classifier.php
@@ -132,6 +132,9 @@ class PageSection_ProfilesClassifier extends Extension_PageSection {
 		
 		try {
 			if(!empty($id) && !empty($do_delete)) { // Delete
+				if(!$active_worker->hasPriv(sprintf("contexts.%s.delete", CerberusContexts::CONTEXT_CLASSIFIER)))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.delete'));
+				
 				DAO_Classifier::delete($id);
 				
 				echo json_encode(array(
@@ -146,9 +149,6 @@ class PageSection_ProfilesClassifier extends Extension_PageSection {
 				@list($owner_context, $owner_context_id) = explode(':', DevblocksPlatform::importGPC($_REQUEST['owner'], 'string', ''));
 				@$do_retrain = DevblocksPlatform::importGPC($_REQUEST['do_retrain'], 'integer', 0);
 				
-				if(empty($name))
-					throw new Exception_DevblocksAjaxValidationError("The 'Name' field is required.", 'name');
-				
 				if(empty($owner_context) || false == Extension_DevblocksContext::get($owner_context))
 					throw new Exception_DevblocksAjaxValidationError("The 'Owner' field is required.");
 				
@@ -156,15 +156,20 @@ class PageSection_ProfilesClassifier extends Extension_PageSection {
 					throw new Exception_DevblocksAjaxValidationError("You don't have permission to use this owner.", 'owner');
 				
 				if(empty($id)) { // New
+					if(!$active_worker->hasPriv(sprintf("contexts.%s.create", CerberusContexts::CONTEXT_CLASSIFIER)))
+						throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.create'));
+					
 					$fields = array(
 						DAO_Classifier::CREATED_AT => time(),
 						DAO_Classifier::UPDATED_AT => time(),
 						DAO_Classifier::NAME => $name,
 						DAO_Classifier::OWNER_CONTEXT => $owner_context,
 						DAO_Classifier::OWNER_CONTEXT_ID => $owner_context_id,
-						DAO_Classifier::DICTIONARY_SIZE => 0,
 						DAO_Classifier::PARAMS_JSON => json_encode([]),
 					);
+					
+					if(!DAO_Classifier::validate($fields, $error))
+						throw new Exception_DevblocksAjaxValidationError($error);
 					
 					if(false == ($id = DAO_Classifier::create($fields)))
 						throw new Exception_DevblocksAjaxValidationError("Failed to create the record.");
@@ -173,12 +178,19 @@ class PageSection_ProfilesClassifier extends Extension_PageSection {
 						C4_AbstractView::setMarqueeContextCreated($view_id, CerberusContexts::CONTEXT_CLASSIFIER, $id);
 					
 				} else { // Edit
+					if(!$active_worker->hasPriv(sprintf("contexts.%s.update", CerberusContexts::CONTEXT_CLASSIFIER)))
+						throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.edit'));
+					
 					$fields = array(
 						DAO_Classifier::UPDATED_AT => time(),
 						DAO_Classifier::NAME => $name,
 						DAO_Classifier::OWNER_CONTEXT => $owner_context,
 						DAO_Classifier::OWNER_CONTEXT_ID => $owner_context_id,
 					);
+					
+					if(!DAO_Classifier::validate($fields, $error, $id))
+						throw new Exception_DevblocksAjaxValidationError($error);
+					
 					DAO_Classifier::update($id, $fields);
 				}
