+++ features/cerberusweb.core/api/uri/profiles/comment.php
@@ -131,6 +131,9 @@ class PageSection_ProfilesComment extends Extension_PageSection {
 				throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translateCapitalized('common.access_denied'));
 			
 			if(!empty($id) && !empty($do_delete)) { // Delete
+				if(!$active_worker->hasPriv(sprintf("contexts.%s.delete", CerberusContexts::CONTEXT_COMMENT)))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.delete'));
+				
 				DAO_Comment::delete($id);
 				
 				echo json_encode(array(
@@ -147,9 +150,6 @@ class PageSection_ProfilesComment extends Extension_PageSection {
 			@$file_ids = DevblocksPlatform::sanitizeArray(DevblocksPlatform::importGPC($_REQUEST['file_ids'],'array',array()), 'int');
 			@$options = DevblocksPlatform::importGPC($_REQUEST['options'],'array',[]);
 			
-			if(empty($comment))
-				throw new Exception_DevblocksAjaxValidationError("The 'Comment' field is required.", 'comment');
-			
 			if(empty($id)) { // New
 				$also_notify_worker_ids = array_keys(CerberusApplication::getWorkersByAtMentionsText($comment));
 				
@@ -165,18 +165,28 @@ class PageSection_ProfilesComment extends Extension_PageSection {
 					DAO_Comment::COMMENT => $comment,
 					DAO_Comment::CREATED => time(),
 				);
+				
+				if(!DAO_Comment::validate($fields, $error))
+					throw new Exception_DevblocksAjaxValidationError($error);
+				
 				$id = DAO_Comment::create($fields, $also_notify_worker_ids, $file_ids);
 				
 				if(!empty($view_id) && !empty($id))
 					C4_AbstractView::setMarqueeContextCreated($view_id, CerberusContexts::CONTEXT_COMMENT, $id);
 				
 			} else { // Edit
+				if(!$active_worker->hasPriv(sprintf("contexts.%s.update", CerberusContexts::CONTEXT_COMMENT)))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.edit'));
+				
 				$fields = array(
 					DAO_Comment::COMMENT => $comment,
 				);
 				
 				if(isset($options['update_timestamp']) && $options['update_timestamp'])
 					$fields[DAO_Comment::CREATED] = time();
+				
+				if(!DAO_Comment::validate($fields, $error, $id))
+					throw new Exception_DevblocksAjaxValidationError($error);
 				
 				DAO_Comment::update($id, $fields);
 			}
